Use C++-accessible Placement in compute_global_tensor_info
This seems to speed things up, though there is more overhead than I'd like to unwrap a Placement via pybind11. Might be time to introduce nanobind; I'm not sure that we can use it in general, but the Placement type system is (currently?) isolated. (Alternatively, I could look into whether we can cut this overhead.)

ghstack-source-id: cccd6111a93c6cb25a814643c49207fb7d609b9d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/163031